Needs to be int

+++ mask2former/modeling/backbone/mrml.py
@@ -306,8 +306,8 @@ class MRML(nn.Module):
         _load_weights(self, checkpoint_path, prefix)
 
     def divide_tokens_to_split_and_keep(self, tokens_at_curr_scale, patches_scale_coords_curr_scale, curr_scale):
-        k_split = tokens_at_curr_scale.shape[1] * self.upscale_ratio
-        k_keep = tokens_at_curr_scale.shape[1] - k_split
+        k_split = int(tokens_at_curr_scale.shape[1] * self.upscale_ratio)
+        k_keep = int(tokens_at_curr_scale.shape[1] - k_split)
         pred_meta_loss = self.metalosses[curr_scale](tokens_at_curr_scale.detach()).squeeze(2)
         tkv, tki = torch.topk(pred_meta_loss, k=k_split, dim=1, sorted=False)
         bkv, bki = torch.topk(pred_meta_loss, k=k_keep, dim=1, sorted=False, largest=False)
